Creates PPD class, and modifies compressor to use it
The PPD class parses a string with the ppd, looking for
its attributes. Now, instead of saving a dictionary with
ppd_path => {size: ppd_size, start: ppd_start}, it creates
a PPD object for each ppd, with it's attributes, so we can
in the future list them as foomatic does.