Archive the PPDs into a tar file
It's better than the sqlite3 solution because we can decompress and load
the tar file from memory, without the need to write to decompress, write
to disk and load, as with sqlite3.

As tar doesn't have an index, it needs to do a sequential search to find
the files. It might be better to use 7z instead.