Generate CHANGES.txt from git log and add it to sdist.
I made git2changes which simply gets the commit messages and organizes
them into something similar to the GNU ChangeLog style. But, instead of
separating the commits by day, I organize them by version. I think this
way, even as it's not standard, generates a file that is more useful to
pyppd's users.