Using cat instead of tarfile for the archive.
Instead of creating a .tar archive, we simply concatenate all PPDs found
in a big text file, and save each's start and length inside of it. Then,
when we have to "cat" it, we decompress the text file and print only
that slice of it.

In my tests, it's around 8% faster than using .tar. Also, it makes it
easier, in the future, to support not only files .ppd, but also .ppd.gz.