Adds support for parsing *.ppd.gz files.
Now, instead of looking for only *.ppd in the directory passed as parameter,
it looks also for *.ppd.gz. When found, it tests: does it ends with
".gz" (case-sensitive (I have to change this))?
If so, opens with gzip. If not, opens as usual. When we archive gzipped PPDs,
we save them uncompressed.