Use PPD's basename instead of "*PCFileName:" as name.
There are some of PPDs which are in different files but contains the
same "*PCFileName:". So, without this change, pyppd overwritten them in
the index, making them impossible to recover. Now, it'll only happen
when there're two PPDs with the same basename. I'm also writing at
ISSUES.txt that we need to figure out how to solve this possible problem.