Fixes bug where it made a new random filename to each PPD entry.
When we have two PPDs with the same filename, we create add a random
number to one of them, so we can add both to our archive. But, if this
PPD which'll have the new name have more than one entry in the list,
each entry will have a different random name. This commit fixes this.