Log only to stderr and use only --verbose (no more --debug).
I don't think logging to a file is worthwile to pyppd, as it isn't a
daemon. If you encounter some problem, you'll be running it yourself,
and not by cron or something. So, you can read in stderr.

Also, I prefer using -v and -vv instead of -v and -d. I think it makes
things more standard.