Consistent byte vs. string handling
Use bytes for reading and writing the archive and PPD files.  Do not try to
decode/encode whole PPDs. Many of them are not in UTF-8, and not even in the
encoding they claim to be in, so it's best to not assume anything about their
encoding at all and just keep them as bytes. Only decode the minimal pieces
like the manufacturer and product names.

This makes pyppd work with Python 3 and is also more robust against encoding
errors.

Avoid calling str() on a PPD object, as that will cause UTF-8 encoding again on
an unicode object. Just call __str__() directly. This is a bit ugly, but works
with both Python 2 and 3.